Updating and first try for unit tests

+++ Solution2_py/database.py
@@ -1,25 +1,37 @@
 import mysql.connector
-import pandas as pd 
 
-def store_in_database(data):
+
+def store_in_database(data, reinit, database_hostname, database_username, database_password, database_name):
     # Connect to the database
     conn = mysql.connector.connect(
-        host="candidaterds.n2g-dev.net",
-        user="cand_62cm",
-        password="3ITMjTgArIDmesgX",
-        database="cand_62cm",
+        host=database_hostname,
+        user=database_username,
+        password=database_password,
+        database=database_name,
     )
-    print(data['timestamp'])
+
     # Create a cursor
     cursor = conn.cursor()
 
-    sql = "DROP TABLE results"
-    cursor.execute(sql)
+    # If you want to re initialize the table:
+    if reinit == True:
+        sql = "DROP TABLE results"
+        cursor.execute(sql)
 
-    sql = "CREATE TABLE IF NOT EXISTS results (gatewayEui BIGINT, profileId INTEGER, endpointId INTEGER, clusterId INTEGER, attributeId INTEGER, timestamp BIGINT, value REAL)"
-    cursor.execute(sql)
+    # Create the table if it does not already exist
+    cursor.execute("""
+        CREATE TABLE IF NOT EXISTS results (
+            gatewayEui BIGINT,
+            profileId INT,
+            endpointId INT,
+            clusterId INT,
+            attributeId INT,
+            timestamp BIGINT,
+            value REAL
+        )
+    """)
 
-    print('oof')
+
     # Insert the data into the table
     sql = "INSERT INTO results (gatewayEui, profileId, endpointId, clusterId, attributeId, timestamp, value) VALUES (%s, %s, %s, %s, %s, %s, %s)"
     # sql = "INSERT INTO results (timestamp, value) VALUES (%s, %s)"
@@ -42,13 +54,13 @@ def store_in_database(data):
     cursor.close()
     conn.close()
 
-def read_from_database():
+def read_from_database(database_hostname, database_username, database_password, database_name):
     # Connect to the database
     conn = mysql.connector.connect(
-        host="candidaterds.n2g-dev.net",
-        user="cand_62cm",
-        password="3ITMjTgArIDmesgX",
-        database="cand_62cm",
+        host=database_hostname,
+        user=database_username,
+        password=database_password,
+        database=database_name,
     )
 
     # Create a cursor
